requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
mysql-connector-python